// types/interfaceUserApi.ts
export interface UserAPI {
  id: number
  first_name: string
  last_name: string
  email: string
  phone: string
  role: string
  state: string
  profile_img: null
  country: string
  city: string
  zip: null
  emergency_contact: string
  emergency_phone: string
  login: boolean
  created_at: Date
  updated_at: Date
}